refactor SQL Mock and add test
