+++ pkg/gofr/datasource/sql/sql.go
@@ -4,14 +4,9 @@ import (
 	"database/sql"
 	"fmt"
 	"strconv"
-	"testing"
 	"time"
 
-	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/XSAM/otelsql"
-	"go.uber.org/mock/gomock"
-	"gofr.dev/pkg/gofr/testutil"
-
 	_ "github.com/lib/pq" // used for concrete implementation of the database driver.
 
 	"gofr.dev/pkg/gofr/config"
@@ -114,21 +109,4 @@ func pushDBMetrics(db *sql.DB, metrics Metrics) {
 
 		time.Sleep(frequency * time.Second)
 	}
-}
-
-func NewSQLMocks(t *testing.T) (*DB, sqlmock.Sqlmock, *MockMetrics) {
-	db, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-	}
-
-	ctrl := gomock.NewController(t)
-	mockMetrics := NewMockMetrics(ctrl)
-
-	return &DB{
-		DB:      db,
-		logger:  testutil.NewMockLogger(testutil.DEBUGLOG),
-		config:  nil,
-		metrics: mockMetrics,
-	}, mock, mockMetrics
 }
